Refactor template method lookup
Refactor template() method from being an instance method to being a
static method. This is so that a component is created for each
usage of the template instead of only a single component instance.
Support was also added for the static template property to be a
string in addition to being a function.